+++ packages/media-query-list-parser/src/nodes/media-feature-boolean.ts
@@ -21,6 +21,11 @@ export class MediaFeatureBoolean {
 		return token[4].value;
 	}
 
+	getNameToken() {
+		const token = (((this.name as TokenNode).value as CSSToken) as TokenIdent);
+		return token;
+	}
+
 	tokens(): Array<CSSToken> {
 		return [
 			...this.before,