+++ packages/media-query-list-parser/test/test.mjs
@@ -65,5 +65,6 @@ import './cases/various/0004.mjs';
 import './cases/various/0005.mjs';
 import './cases/various/0006.mjs';
 import './cases/various/0007.mjs';
+import './cases/various/0008.mjs';
 
 import './serialize/0001.mjs';