+++ packages/media-query-list-parser/CHANGELOG.md
@@ -1,3 +1,7 @@
+### Unreleased
+
+- Add `getName` and `getNameToken` to all nodes that have a feature name.
+
 ### 1.0.0 (November 14, 2022)
 
 - Initial version